refactor: moved DashboardRentalsController and DashboardRentedController to own directory
