+++ app/Http/Controllers/Dashboard/DashboardRentedController.php
@@ -1,7 +1,8 @@
 <?php
 
-namespace App\Http\Controllers;
+namespace App\Http\Controllers\Dashboard;
 
+use App\Http\Controllers\Controller;
 use Inertia\Inertia;
 
 class DashboardRentedController extends Controller
