+++ routes/web.php
@@ -1,8 +1,9 @@
 <?php
 
 use App\Http\Controllers\ApartmentController;
-use App\Http\Controllers\DashboardRentalsController;
-use App\Http\Controllers\DashboardRentedController;
+use App\Http\Controllers\Dashboard\DashboardMarkedController;
+use App\Http\Controllers\Dashboard\DashboardRentalsController;
+use App\Http\Controllers\Dashboard\DashboardRentedController;
 use App\Http\Controllers\MarkController;
 use App\Http\Controllers\RentController;
 use App\Http\Controllers\ReviewController;
@@ -24,6 +25,11 @@ Route::resource('apartments', ApartmentController::class)->except(['index']);
 Route::apiResource('apartments.reviews', ReviewController::class)->shallow();
 Route::apiResource('apartments.rents', RentController::class);
 Route::apiResource('apartments.marks', MarkController::class)->only(['store', 'destroy'])->shallow();
+
+Route::prefix('dashboard')->name('dashboard.')->group(function () {
+    Route::get('rentals', DashboardRentalsController::class)->name('rentals');
+    Route::get('rented', DashboardRentedController::class)->name('rented');
+});
 
 Route::get('dashboard/rentals', DashboardRentalsController::class)
     ->middleware(['auth', 'verified'])
